--- utils/SmartPyBasic/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "SmartPyBasic",
-  "version": "1.0.0",
-  "description": "",
-  "main": "smartml-cli.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "bs58check": "^2.1.2",
-    "chalk": "^4.1.0",
-    "libsodium-wrappers-sumo": "^0.7.6"
-  }
-}
