--- utils/SmartPyBasic/version.py
@@ -1,1 +1,0 @@
-version = "dev"
