+++ utils/scripts/compileCode.sh
@@ -1,1 +1,1 @@
-./utils/SmartPyBasic/SmartPy.sh compile ./contract/demo.py "MyContract(12, 123)" ./contract_build
+./utils/smartpy-cli/SmartPy.sh compile ./contract/demo.py "MyContract(12, 123)" ./contract_build
