+++ utils/smartpy-cli/version.py
@@ -1,0 +1,1 @@
+version = "dev"
